Implement Point 1.2
This is the first commit that address's point 1.2 which states that the
Add method should be able to handle 1 or 1 numbers provided.

I have refactored the Add method to test for Null, Empty or Whitespace
filled instances of the numbers parameter.  I decided to add tests to
check for null and whitepace filled numbers parameters.

I have then added a single test that checks for the result to match the
value of numbers when passed to the Add method with a single number.