Implement Point 4
This commit make the changes required to pass point 4 of the kata.

The changes here allow the delimiter to be passed as part of the numbers
parameter.  Here you can see the evolution of the way the delimiter data
is generated from the original char constant, through being a char array
of acceptable values to being read from the input.