Implement Point 7
This commit adds the ability to handle variable length delimiters.  I have refactored the GetDelimiterFrom(string) method to handle the new logic and added a new test to drive these changes.  In reality, the refactoring stage of this test would probably see me refactoring the code to use a Strategy based solution to remove the conditional based logic.